services/timesrv: implement clock interfaces properly
